+++ SCM.UI/Models/DTOs/Products/ProductDTO.cs
@@ -2,8 +2,8 @@
 {
     public class ProductDTO
     {
-        public int Id { get; set; }
-        public int CategoryId { get; set; }
+        public long Id { get; set; }
+        public long CategoryId { get; set; }
         public string Name { get; set; }
         public string Detail { get; set; }
         public int UnitInStock { get; set; }
