+++ SCM.UI/Models/RequestModels/Approves/ApproveVM.cs
@@ -2,7 +2,7 @@
 {
     public class ApproveVM
     {
-        public int RequestId { get; set; }
+        public long RequestId { get; set; }
         public bool IsApproved { get; set; }
     }
 }
