+++ SCM.UI/Models/RequestModels/Offers/CreateOfferVM.cs
@@ -2,7 +2,7 @@
 {
     public class CreateOfferVM
     {
-        public int RequestId { get; set; }
+        public long RequestId { get; set; }
         public string SupplierName { get; set; }
 
         public decimal Amount { get; set; }
