+++ SCM.UI/Models/RequestModels/Products/UpdateProductVM.cs
@@ -2,11 +2,11 @@
 {
     public class UpdateProductVM
     {
-        public int Id { get; set; }
-        public int? CategoryId { get; set; }
+        public long Id { get; set; }
+        public long? CategoryId { get; set; }
         public string Name { get; set; }
         public string Detail { get; set; }
-        public int? UnitInStock { get; set; }
+        public long? UnitInStock { get; set; }
         public decimal? UnitPrice { get; set; }
     }
 }
